Project is build able, and data is send to the server.

+++ PlayTestToolkit/Packages/PlayTestToolkit/Runtime/DataRecorders/InitialData.cs
@@ -2,6 +2,6 @@
 {
     public class InitialData
     {
-        public int startTime;
+        public double startTime;
     }
 }
